Services/ChatService/requirements.txt
fastapi==0.103.1
python-dotenv==1.0.0
SQLAlchemy==2.0.20
uvicorn==0.23.2
psycopg2==2.9.7
asyncpg==0.28.0
pydantic-settings==2.0.3
websockets==11.0.3
